+++ dify_studio/opus_tag_smith/nodes/formatter.py
@@ -1,5 +1,6 @@
 # pylint: disable=missing-module-docstring
 
+# todo write unit test for OTS formatter
 
 import re
 
@@ -11,8 +12,9 @@ SEP = ","
 # extract keys  ################################################################
 
 
-EXTRACT_TITLE_KEY = "title"
 EXTRACT_YEAR_KEY = "release_year"
+EXTRACT_TITLE_KEY = "title"
+EXTRACT_SUBTITLE_KEY = "subtitle"
 EXTRACT_TAGS_KEY = "tags"
 
 # opus extract
@@ -64,7 +66,14 @@ def _add_party_entries(parties, prefix):
 # formatter  ###################################################################
 
 
-def _format_opus(extract, title, year, tags):  # ===============================
+def _format_opus(extract):  # ==================================================
+    year = extract[EXTRACT_YEAR_KEY]
+
+    # title  -------------------------------------------------------------------
+    title = _convert_filename_safe(extract[EXTRACT_TITLE_KEY])
+    subtitle = _convert_filename_safe(extract[EXTRACT_SUBTITLE_KEY])
+
+    # fixme utilize subtitle for OTS Opus
 
     # season & episode  --------------------------------------------------------
     season_and_episode = []
@@ -89,43 +98,51 @@ def _format_opus(extract, title, year, tags):  # ===============================
         season_and_episode_content = "".join(season_and_episode)
         title = title + "." + season_and_episode_content
 
+    tags = extract[EXTRACT_TAGS_KEY]
+
     # title names  -------------------------------------------------------------
     safe_title = _convert_filename_safe(title)
     folder_name = "[{year}]{title}".format(year=year, title=safe_title)
     resource_name = folder_name + "{" + SEP.join(tags) + "}"
 
     # create response  ---------------------------------------------------------
-    if title == safe_title:
-        title_part = "Title:\n```\n{}\n```\n".format(title)
-    else:
-        title_part = (
-            "Title:\n```\n{}\n```\nTitle (Safe):\n```\n{}\n```\n".format(
-                title, safe_title
-            )
-        )
-
-    sne_part = ""
-    if season_and_episode_content:
-        sne_part = """Episode Name:
+    response = """
+Title:
 
 ```
-{}
+{title}
 ```
-""".format(season_and_episode_content)
 
-    response = title_part + sne_part + """
+Title (Safe):
+
+```
+{safe_title}
+```
+
+Episode Name:
+
+```
+{sne}
+```
+
 Folder Name:
 
 ```
-{}
+{folder_name}
 ```
 
 Resource Name:
 
 ```
-{}
+{resource_name}
 ```
-""".format(folder_name, resource_name)
+""".format(
+        title=title,
+        safe_title=safe_title,
+        sne=season_and_episode_content,
+        folder_name=folder_name,
+        resource_name=resource_name,
+    )
 
     return response
 
@@ -221,7 +238,7 @@ def main(extract: dict, target: str):
     tags = extract[EXTRACT_TAGS_KEY]
 
     response = (
-        _format_opus(extract, title, year, tags)
+        _format_opus(extract)
         if target == "Opus"
         else _format_shelver(extract, title, year, tags)
     )